09-27-2017

+++ GenericBackEnd/ICursiVisionBackEnd.cpp
@@ -14,7 +14,7 @@ SetProperties:
 
       IUnknown *pIUnknown = NULL;
       QueryInterface(IID_IUnknown,reinterpret_cast<void **>(&pIUnknown));
-      pIGProperties -> ShowProperties((long)hwndParent,pIUnknown);
+      pIGProperties -> ShowProperties(hwndParent,pIUnknown);
       pIUnknown -> Release();
 
    }
